Transition Treehouse icon to cutout
This is going to sound so crazy but I didn't make this icon the right way. You can totally see how it's wrong in the dark mode of the site. That's right, I said ~dark mode~ boo. Basically, the other two icons are just hollow inside which makes it super easy to invert them. This was not so here with the Treehouse icon honey! It's all good now though, I fired up Illustrator and righted this wrong!